+++ torch_frame/gbdt/gbdt.py
@@ -1,6 +1,7 @@
+from __future__ import annotations
+
 import os
 from abc import abstractmethod
-from typing import Optional
 
 import torch
 from torch import Tensor
@@ -32,8 +33,8 @@ class GBDT:
     def __init__(
         self,
         task_type: TaskType,
-        num_classes: Optional[int] = None,
-        metric: Optional[Metric] = None,
+        num_classes: int | None = None,
+        metric: Metric | None = None,
     ):
         self.task_type = task_type
         self._is_fitted: bool = False