+++ torch_frame/nn/conv/table_conv.py
@@ -22,6 +22,5 @@ class TableConv(Module, ABC):
         """
         raise NotImplementedError
 
-    def reset_parameters(self):
+    def reset_parameters(self) -> None:
         r"""Resets all learnable parameters of the module."""
-        pass